+++ pygpsclient/menu_bar.py
@@ -9,11 +9,10 @@ Created on 12 Sep 2020
 :copyright: SEMU Consulting © 2020
 :license: BSD 3-Clause
 """
-# pylint: disable=invalid-name, too-many-instance-attributes, too-many-ancestors
 
 from tkinter import Menu, DISABLED
 
-from .strings import (
+from pygpsclient.strings import (
     MENUFILE,
     MENUVIEW,
     MENUOPTION,